Fix config, fix log path, rename dbcontext, add FilesystemBlobStorage, add web appsettings, add dockerfile, add RemovePlayerHandler (currently does not function)

+++ src/Skateboard3Server.Web/Services/Models/SkateProfile/SetUserAchievements.cs
@@ -5,7 +5,7 @@ namespace Skateboard3Server.Web.Services.Models.SkateProfile
     public class SetUserAchievements
     {
         public PlatformType PlatformId { get; set; }
-        public long UserId { get; set; }
+        public uint UserId { get; set; }
         public string AchievementIds { get; set; }
     }
 }
