+++ src/Skateboard3Server.Web/Services/Models/SkateProfile/SetUserDlc.cs
@@ -5,7 +5,7 @@ namespace Skateboard3Server.Web.Services.Models.SkateProfile
     public class SetUserDlc
     {
         public PlatformType PlatformId { get; set; }
-        public long UserId { get; set; }
+        public uint UserId { get; set; }
         public string OfferIds { get; set; }
     }
 }
