+++ src/Skateboard3Server.Web/Services/Models/SkateProfile/StartLoginProcess.cs
@@ -5,7 +5,7 @@ namespace Skateboard3Server.Web.Services.Models.SkateProfile
     public class StartLoginProcess
     {
         public PlatformType PlatformType { get; set; }
-        public long UserId { get; set; }
+        public uint UserId { get; set; }
         public string SessionKey { get; set; }
         public bool DeleteThumbs { get; set; }
         public int Difficulty { get; set; } //TODO: enum?
